fixes the tests
